fix: A few more fixes

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/server/tileentity/DragonBeaconTileEntity.java
@@ -23,7 +23,7 @@ import java.util.Random;
 
 public class DragonBeaconTileEntity extends BaseBlockTileEntity{
 	public Type type = Type.NONE;
-	public int tick;
+	public float tick;
 
 	public final float bobOffs;
